requirements.txt
livekit-agents==1.1.4
livekit-api==0.17.0
livekit-plugins-openai==0.5.0
livekit-plugins-silero==0.5.0
groq>=0.9.0
elevenlabs>=1.0.0
openai>=1.40.0
python-dotenv>=1.0.0
loguru>=0.7.0
asyncio>=3.4.3
aiohttp>=3.8.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
numpy>=1.24.0
